requirements.txt
aiogram==2.25.1
APScheduler==3.11.0
gTTS==2.5.4
python-dotenv==1.0.0
